+++ core/src/main/java/org/svip/repair/fix/Fix.java
@@ -8,9 +8,11 @@ package org.svip.repair.fix;
  * @param fixed change to make
  * @author Juan Francisco Patino
  */
-public record Fix<T>(T old, T fixed) {
+public record Fix<T> (T old, T fixed) {
+
     @Override
     public String toString() {
         return old + " -> " + fixed;
     }
+
 }